+++ src/main/java/com/commercetools/sync/products/ProductSyncOptionsBuilder.java
@@ -86,7 +86,6 @@ public final class ProductSyncOptionsBuilder
             errorCallback,
             warningCallback,
             batchSize,
-            removeOtherLocales,
             removeOtherSetEntries,
             removeOtherCollectionEntries,
             removeOtherProperties,